WIP: optimize all class files
