+++ src/formulize.plugin.d.ts
@@ -1,0 +1,7 @@
+import { FormulizePlugin } from './formulize.interface';
+
+declare global {
+    interface JQuery {
+        formulize: FormulizePlugin;
+    }
+}
